JAIZelAnime: Match leftover startAnimSound (USA partial)

+++ src/JAZelAudio/JAIZelAnime.cpp
@@ -17,7 +17,6 @@ void JAIZelAnime::setAnimSound(Vec* pos, f32 frame, f32 rate, u32 mtrlSndId, s8 
 }
 
 /* 802AC888-802ACD34       .text startAnimSound__11JAIZelAnimeFPvUlPP8JAISoundPQ27JAInter5ActorUc */
-// NONMATCHING - 1 missing reg move
 void JAIZelAnime::startAnimSound(void* i_basic, u32 i_soundID, JAISound** i_sound, JAInter::Actor* i_actor, u8 param_4) {
     JAIZelBasic* basic = (JAIZelBasic*)i_basic;
 
@@ -88,7 +87,7 @@ void JAIZelAnime::startAnimSound(void* i_basic, u32 i_soundID, JAISound** i_soun
     }
 
     u32 actor_c = i_actor->field_0xc;
-    int var_r27 = actor_c >> 24;
+    u16 var_r27 = actor_c >> 24;
     i_actor->field_0xc = actor_c & 0xFFFFFF;
 
     if (i_actor->field_0x4 != NULL) {
